Edit #4 - Ammo and Health Customisation
In this edit, I installed another variation of my beginning font and
then used this in my ammo and health panels. I also removed all of the
components of the health panel other than the player health value. I
also removed the background of the ammo panel and changed the font size.
Also added was a line below the player health value to show the player
what team they are on. This is my biggest and most complex update so
far!